python/requirements.txt
fastapi == 0.115.12
uvicorn[standard] == 0.34.0
python-multipart == 0.0.20
ffmpeg-python == 0.2.0